Restore stable version of reconciliation page and update notifications
Rollback BLReconciliation component to a previous stable state and switch to standard shadcn/ui Toaster, resolving React error #310.
