+++ ROLLBACK-SOLUTION-REACT-ERROR-310.md
@@ -1,0 +1,96 @@
+# Rollback Solution - React Error #310
+
+## ✅ Problème Identifié
+L'erreur React #310 en production est apparue **après l'ajout des actions sur les lignes et des 2 onglets**. La version précédente (backup) fonctionnait parfaitement.
+
+## 🔄 Action Effectuée
+**Rollback vers la version fonctionnelle :**
+- `BLReconciliation_backup.tsx` → `BLReconciliation.tsx` (restauré)
+- `BLReconciliation.tsx` récente → `BLReconciliation_recent_broken.tsx` (sauvegardé)
+- Restauration du système de notifications shadcn/ui standard
+
+## 📊 Comparaison des Versions
+
+### Version Backup (Fonctionnelle) - 21KB
+✅ **Caractéristiques :**
+- Interface simple avec 2 onglets (Manuel/Automatique)
+- Actions limitées : Voir détails + Dévalider automatique uniquement
+- Pas de pagination complexe
+- Une seule fonction d'action : `handleDevalidateAutoReconciliation`
+- shadcn/ui standard (Button, Tabs, Badge, Input)
+
+### Version Récente (Cassée) - 35KB  
+❌ **Problèmes introduits :**
+- Actions complexes : Valider/Dévalider/Supprimer
+- Système de pagination complet avec mutations multiples
+- Composants UI personnalisés (RobustButton, RobustInput, etc.)
+- Système de permissions complexe
+- Gestion d'états multiples
+
+## 🎯 Fonctionnalités de la Version Restaurée
+
+### Interface
+- ✅ Onglets Manuel vs Automatique
+- ✅ Recherche par fournisseur/BL/facture
+- ✅ Affichage des livraisons avec statut "delivered"
+- ✅ Design cohérent shadcn/ui
+
+### Actions Disponibles
+- ✅ **Voir les détails** (icône œil)
+- ✅ **Dévalider rapprochement automatique** (directeurs/admins uniquement)
+- ✅ Filtrage automatique selon mode fournisseur
+- ✅ Notifications toast
+
+### Permissions
+- ✅ Accès restreint : managers, directeurs, admins uniquement
+- ✅ Actions de dévalidation : directeurs et admins uniquement
+- ✅ Redirection automatique pour employés
+
+## 🚀 Plan de Réimplémentation Progressive
+
+### Phase 1 : Test de la Version Backup ✅
+- [ ] Déployer la version backup en production
+- [ ] Vérifier absence d'erreur React #310
+- [ ] Confirmer fonctionnement des onglets
+- [ ] Tester l'action de dévalidation
+
+### Phase 2 : Ajout Progressif des Fonctionnalités
+1. **Pagination simple** (si nécessaire)
+2. **Action Valider** (une à la fois)
+3. **Action Supprimer** (après validation des précédentes)
+4. **Améliorations UI** (en dernier)
+
+### Phase 3 : Test à Chaque Étape
+- Test en développement
+- Test en production
+- Rollback immédiat si erreur React #310
+
+## 📁 Fichiers de Sauvegarde
+
+```
+client/src/pages/BLReconciliation.tsx              # ✅ Version backup restaurée
+client/src/pages/BLReconciliation_recent_broken.tsx # ❌ Version avec erreur React #310
+client/src/pages/BLReconciliation_shadcn_broken.tsx # ❌ Autre version problématique
+client/src/hooks/use-toast-robust.ts               # 🔧 Composants robustes (inutilisés)
+client/src/components/ToasterRobust.tsx            # 🔧 Composants robustes (inutilisés)
+```
+
+## 🔍 Points de Test Production
+
+**À vérifier immédiatement :**
+- [ ] Page `/bl-reconciliation` se charge sans erreur React #310
+- [ ] Onglets Manuel/Automatique switchent correctement
+- [ ] Recherche fonctionne
+- [ ] Action "Dévalider automatique" fonctionne (directeur/admin)
+- [ ] Notifications s'affichent
+- [ ] Autres modules restent inchangés
+
+## 💡 Conclusion
+
+Cette approche de **rollback progressif** permet de :
+1. **Éliminer immédiatement** l'erreur React #310
+2. **Conserver les fonctionnalités essentielles** du module
+3. **Réajouter les features** de manière contrôlée
+4. **Identifier précisément** quelle modification cause l'erreur
+
+**Status : ✅ Version backup restaurée - Prêt pour test production**
