+++ client/src/App.tsx
@@ -1,6 +1,6 @@
 import { queryClient } from "./lib/queryClient";
 import { QueryClientProvider } from "@tanstack/react-query";
-import { ToasterRobust } from "@/components/ToasterRobust";
+import { Toaster } from "@/components/ui/toaster";
 import { TooltipProvider } from "@/components/ui/tooltip";
 import RouterProduction from "@/components/RouterProduction";
 import ErrorBoundary from "@/components/ErrorBoundary";
@@ -10,7 +10,7 @@ function App() {
     <ErrorBoundary>
       <QueryClientProvider client={queryClient}>
         <TooltipProvider>
-          <ToasterRobust />
+          <Toaster />
           <RouterProduction />
         </TooltipProvider>
       </QueryClientProvider>
